+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Fix #7538: bump cert-manager from 1.18.2 to 1.19.4
 
 #### New Features
+* Fix #7451: Add support for server-side content negotiation (Table and PartialObjectMetadata responses) 
 
 #### _**Note**_: Breaking changes
 * Fix #7544: cluster-api model classes moved from package `io.fabric8.kubernetes.api.model.clusterapi.v1beta1` to `io.fabric8.kubernetes.api.model.clusterapi.core.v1beta1` (following upstream cluster-api v1.11+ API reorganization)